requirements.txt
openai>=1.12.0
azure-search-documents
azure-identity
azure-storage-blob
python-dotenv
streamlit
tiktoken
